WA-238 DailyLimit Record types
